修改 decimal 代替 double

+++ csharp/ToolGood.Algorithm2.Test/AlgorithmEngine/Cylinder.cs
@@ -39,7 +39,7 @@ namespace ToolGood.Algorithm
             {
                 if (operands.Count == 1)
                 {
-                    var r = operands[0].ToNumber().NumberValue;
+                    var r = (double)operands[0].ToNumber().NumberValue;
                     return r * r * Math.PI;
                 }
             }
