+++ csharp/ToolGood.Algorithm2/AlgorithmEngineEx.cs
@@ -626,7 +626,7 @@ namespace ToolGood.Algorithm
                     LastError = obj.ErrorMsg;
                     return def;
                 }
-                return obj.NumberValue;
+                return (double)obj.NumberValue;
             } catch (Exception ex) {
                 LastError = ex.Message + "\r\n" + ex.StackTrace;
             }
@@ -927,7 +927,7 @@ namespace ToolGood.Algorithm
                     LastError = obj.ErrorMsg;
                     return def;
                 }
-                return obj.NumberValue;
+                return (double)obj.NumberValue;
             } catch (Exception ex) {
                 LastError = ex.Message + "\r\n" + ex.StackTrace;
             }
