+++ csharp/ToolGood.Algorithm2/LitJson/JsonData.cs
@@ -10,7 +10,7 @@ namespace ToolGood.Algorithm.LitJson
         #region Fields
         private IList<JsonData> inst_array;
         private bool inst_boolean;
-        private double inst_double;
+        private decimal inst_double;
         internal IDictionary<string, JsonData> inst_object;
         private string inst_string;
         private JsonType type;
@@ -72,10 +72,10 @@ namespace ToolGood.Algorithm.LitJson
             inst_boolean = val;
         }
 
-        void IJsonWrapper.SetDouble(double val)
+        void IJsonWrapper.SetDouble(decimal val)
         {
             type = JsonType.Double;
-            inst_double = val;
+            inst_double = (decimal)val;
         }
 
         void IJsonWrapper.SetString(string val)
@@ -154,7 +154,7 @@ namespace ToolGood.Algorithm.LitJson
                     break;
 
                 case JsonType.Double:
-                    inst_double = default(Double);
+                    inst_double = default(decimal);
                     break;
 
                 case JsonType.Boolean:
@@ -172,7 +172,7 @@ namespace ToolGood.Algorithm.LitJson
 
 
         public bool BooleanValue { get { return inst_boolean; } }
-        public double NumberValue { get { return inst_double; } }
+        public decimal NumberValue { get { return inst_double; } }
         public string StringValue { get { return inst_string; } }
     }
 }
